Resolved dropdown position issue on Android
