+++ example/ios/Podfile
@@ -12,8 +12,9 @@ target 'ReactNativeUiComponentsExample' do
   #
   # Note that if you have use_frameworks! enabled, Flipper will not work and
   # you should disable these next few lines.  
-  # use_flipper!({ "Flipper-DoubleConversion" => "1.1.7" }) # to fix build error 'duplicate symbol 'folly::detail::str_to_bool(folly::Range<char const*>*)' in:'
-  # post_install do |installer|
-  #   flipper_post_install(installer)
-  # end
+  use_flipper!({ "Flipper-DoubleConversion" => "1.1.7" }) # to fix build error 'duplicate symbol 'folly::detail::str_to_bool(folly::Range<char const*>*)' in:'
+  
+  post_install do |installer|
+    flipper_post_install(installer)
+  end
 end
