+++ example/package.json
@@ -9,10 +9,15 @@
     "start": "react-native start"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.0.10",
+    "@react-navigation/native-stack": "^6.6.2",
     "react": "17.0.2",
     "react-native": "0.66.4",
+    "react-native-gesture-handler": "^2.4.2",
     "react-native-paper": "^4.0.1",
     "react-native-reanimated": "^2.3.1",
+    "react-native-safe-area-context": "^4.2.5",
+    "react-native-screens": "^3.13.1",
     "react-native-vector-icons": "^9.0.0"
   },
   "devDependencies": {
